requirements.txt
python-dotenv
requests
langchain-core
langchain-mcp
langchain-mcp-adapters
langchain-openai
langgraph